Complete Extra Exercise

+++ module2/src/oop_collection_exception_regex_review/manager_vehicle_save_to_file/models/Truck.java
@@ -1,9 +1,9 @@
 package oop_collection_exception_regex_review.manager_vehicle_save_to_file.models;
 
 public class Truck extends Vehicle {
-    private int maximumLoad;
+    private double maximumLoad;
 
-    public Truck(String licensePlate, Manufacturer manufacturer, int productionYear, String owner, int maximumLoad) {
+    public Truck(String licensePlate, Manufacturer manufacturer, int productionYear, String owner, double maximumLoad) {
         super(licensePlate, manufacturer, productionYear, owner);
         this.maximumLoad = maximumLoad;
     }
@@ -12,11 +12,11 @@ public class Truck extends Vehicle {
 
     }
 
-    public int getMaximumLoad() {
+    public double getMaximumLoad() {
         return maximumLoad;
     }
 
-    public void setMaximumLoad(int maximumLoad) {
+    public void setMaximumLoad(double maximumLoad) {
         this.maximumLoad = maximumLoad;
     }
